+++ src/runners/singleCommandRunner.ts
@@ -11,6 +11,8 @@ export const runSingleCommand = vscode.commands.registerCommand('extension.mateo
     logger.info(vscode.window.activeTextEditor?.document.fileName || "");
 
     let code = codeUtils.getSelectedCode();
+    code = deleteLastMultiLineBreak(code);
+
     if (code === '') {
         return;
     }
@@ -22,12 +24,116 @@ export const runSingleCommand = vscode.commands.registerCommand('extension.mateo
 
     let baseDir: string | undefined = vscode.window.activeTextEditor?.document.fileName;
 
-    let codeSplit = code.split(/(?:\r)?\n/);
+    let codeSplit = splitLines(code);
     //vscode.window.showInformationMessage("Running '" + fileName + "' on: " + config.mateoHostUrl)
 
     await executeLineByLine(codeSplit, baseDir);
-
 });
+
+function splitLines(code: string): string[] {
+    let START_MULTILINE_PATTERN = new RegExp(/\s*=\s*\"{3}\s*$/);
+    let multilineFound: boolean = false;
+    let codeSplit = code.split(/(?:\r)?\n/);
+    let result: string[] = [];
+    let multiLineStartLevel: number = 0;
+    codeSplit = removeBackslashLineBreaks(codeSplit);
+    for (let i = 0; i < codeSplit.length; i++) {
+        if (multilineFound) {
+            if (isEndOfMultiLineString(multiLineStartLevel, codeSplit[i])) {
+                result.push(deleteLastMultiLineBreak(result.pop()!));
+                codeSplit[i] = codeSplit[i].trimStart().replaceAll(/"""/g, '"');
+                multilineFound = false;
+            } else if (isInterimEndOfMultiLineString(multiLineStartLevel, codeSplit[i])) {
+                result.push(deleteLastMultiLineBreak(result.pop()!));
+                codeSplit[i] = codeSplit[i].trim().replaceAll(/"""/g, '"');
+            } else {
+                codeSplit[i] = codeSplit[i] + "\n";
+            }
+            codeSplit[i] = result.pop() + codeSplit[i];
+        }
+        if (START_MULTILINE_PATTERN.test(codeSplit[i])) {
+            multiLineStartLevel = getLevelFromLine(codeSplit[i]);
+            codeSplit[i] = codeSplit[i].trimEnd().replaceAll(/"""/g, '"');
+            multilineFound = true;
+        }
+        result.push(codeSplit[i]);
+    }
+    return result;
+}
+
+function deleteLastMultiLineBreak(line: string):string{
+    let indexOfLastMultiLineBreak = line.lastIndexOf("\n");
+    if (indexOfLastMultiLineBreak >= 0){
+        return line.substring(0,indexOfLastMultiLineBreak);
+    }else{
+        return line;
+    }
+}
+
+function removeBackslashLineBreaks(lines: string[]): string[] {
+    let result: string[] = [];
+    for (let i = 0; i < lines.length; i++) {
+        if (lines[i].endsWith("\\")) {
+            lines[i] = lines[i].substring(0, lines[i].length - 1);
+            lines[i + 1] = lines[i] + lines[i + 1];
+        } else {
+            result.push(lines[i]);
+        }
+    }
+    return result;
+}
+
+function isEndOfMultiLineString(multiLineStartLevel: number, line: string): boolean {
+    let END_COMMAND_MULTILINE_PATTERN = new RegExp(/\s*\"{3}.*\):?\s*$/);
+    let LINE_BREAK_MULTILINE_PATTERN = new RegExp(/\s*\"{3}(.*)(\\\s*)$/);
+    return multiLineStartLevel == getLevelFromLine(line) && (
+        END_COMMAND_MULTILINE_PATTERN.test(line) || LINE_BREAK_MULTILINE_PATTERN.test(line))
+}
+
+function isInterimEndOfMultiLineString(multiLineStartLevel: number, line: string): boolean {
+    let INTERIM_END_MULTILINE_PATTERN = RegExp(/\s*\"{3}.*/);
+    return multiLineStartLevel == getLevelFromLine(line) && INTERIM_END_MULTILINE_PATTERN.test(line);
+}
+
+function getLevelFromLine(line: string): number {
+    let TAB_WIDTH = 4;
+    let level: number = 0;
+    let leave: boolean = false;
+    for (let i = 0; i < line.length; i++) {
+        switch (line.charCodeAt(i)) {
+            case 9:
+                level = level + TAB_WIDTH;
+                break;
+            case 32:
+            case 160:
+            case 8192:
+            case 8193:
+            case 8194:
+            case 8195:
+            case 8196:
+            case 8197:
+            case 8198:
+            case 8199:
+            case 8120:
+            case 8201:
+            case 8239:
+            case 8287:
+            case 12288:
+                level = level + 1;
+                break;
+            default:
+                leave = true;
+                break;
+        }
+        if (leave) {
+            break;
+        }
+    }
+    if (!leave) {
+        level = 0;
+    }
+    return level / TAB_WIDTH + (level % TAB_WIDTH > 0 ? 1 : 0);
+}
 
 async function executeLineByLine(lines: Array<string> | undefined, baseDir: string | undefined) {
     if (lines === undefined || lines.length == 0) {
@@ -36,7 +142,7 @@ async function executeLineByLine(lines: Array<string> | undefined, baseDir: stri
 
     let codeLine = lines[0];
 
-    let COMMAND_PATTERN: string = "^(\\s*)([a-z][a-zA-Z0-9\\s\\-_]*)\\((.*)\\)\\s*";
+    let COMMAND_PATTERN: string = "^(\\s*)([a-z][a-zA-Z0-9\\s\\-_]*)\\(((.|\n)*)\\)\\s*";
     let STEP_PATTERN: string = "^(\\s*)([A-Z][a-zA-Z0-9\\s\\-_]*)\\((.*)\\)\\s*";
     let REF_STEP_PATTERN: string = "^(\\s*)([a-zA-Z0-9\\s\\-_]*)\\.([A-Z][a-zA-Z0-9\\s\\-_]*)\\((.*)\\)\\s*";
     let PARAMETER_VALUES_PATTERN = /,(?=([^"]*"[^"]*")*[^"]*$)/;
@@ -72,7 +178,7 @@ async function executeLineByLine(lines: Array<string> | undefined, baseDir: stri
         }
 
         logger.info("Executing: " + JSON.stringify(commandObject))
-        
+
         let filenameEncoded = encodeURIComponent(baseDir + "");
 
         axios.post(config.mateoHostUrl + '/command?fileName=' + filenameEncoded, commandObject, queryParams)